rfc(dashboard): refactor Setting page, move theme card to single template
